+++ app/generator/CHANGELOG.md
@@ -19,6 +19,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ---
 
+## [0.13.0] - 2025-07-??
+
+### Added
+
+### Changed
+
+- Refactored tests for elimination of "base" directory
+- Modified generator to combine all generated complex datatypes into a single file to resolve circular references
+
+### Deprecated
+
+### Removed
+
+- Eliminated use of "base" directory for parsable-xxx-map files
+
+### Fixed
+
+### Security
+
+
 ## [0.12.0] - 2025-07-12
 
 ### Changed